renamed preset dsahboard to event
