+++ app/src/main/java/com/example/mini_pekkas/MainActivity.java
@@ -26,7 +26,7 @@ public class MainActivity extends AppCompatActivity {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_notifications, R.id.navigation_camera, R.id.navigation_profile)
+                R.id.navigation_home, R.id.navigation_event, R.id.navigation_notifications, R.id.navigation_camera, R.id.navigation_profile)
                 .build();
 
         //NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_activity_main);
