+++ app/src/main/java/com/example/mini_pekkas/ui/event/EventFragment.java
@@ -1,4 +1,4 @@
-package com.example.mini_pekkas.ui.dashboard;
+package com.example.mini_pekkas.ui.event;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,22 +10,22 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.example.mini_pekkas.databinding.FragmentDashboardBinding;
+import com.example.mini_pekkas.databinding.FragmentNotificationsBinding;
 
-public class DashboardFragment extends Fragment {
+public class EventFragment extends Fragment {
 
-    private FragmentDashboardBinding binding;
+    private FragmentNotificationsBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        DashboardViewModel dashboardViewModel =
-                new ViewModelProvider(this).get(DashboardViewModel.class);
+        EventViewModel eventViewModel =
+                new ViewModelProvider(this).get(EventViewModel.class);
 
-        binding = FragmentDashboardBinding.inflate(inflater, container, false);
+        binding = FragmentNotificationsBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textDashboard;
-        dashboardViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        final TextView textView = binding.textNotifications;
+        eventViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
