+++ app/src/main/java/com/example/mini_pekkas/ui/event/EventViewModel.java
@@ -1,16 +1,16 @@
-package com.example.mini_pekkas.ui.dashboard;
+package com.example.mini_pekkas.ui.event;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class EventViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public DashboardViewModel() {
+    public EventViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is dashboard fragment");
+        mText.setValue("This is event fragment");
     }
 
     public LiveData<String> getText() {
