+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">mini-pekkas</string>
     <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
+    <string name="title_event">Event</string>
     <string name="title_notifications">Notifications</string>
     <string name="title_camera">Camera</string>
     <string name="title_profile">Profile</string>
